pyproject.toml
[project]
name = "fastapi-backend-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.5",
    "bcrypt>=3.2.0,<4.0.0",
    "celery>=5.5.3",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.116.1",
    "opentelemetry-api>=1.36.0",
    "opentelemetry-exporter-otlp>=1.36.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.36.0",
    "opentelemetry-instrumentation-fastapi>=0.57b0",
    "opentelemetry-instrumentation-logging>=0.57b0",
    "opentelemetry-sdk>=1.36.0",
    "orjson>=3.11.3",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "redis-data-structures>=0.1.26",
    "sqlalchemy>=2.0.43",
]

[dependency-groups]
dev = ["mypy>=1.17.1", "pre-commit>=4.3.0", "ruff>=0.12.12"]
